Cambiado tipo de aptitud a double

+++ Practica 1/lib/Instance.hpp
@@ -59,7 +59,7 @@ class Instance {
 		}
 
 		//Funcion virtual que calcula el valor de aptitud de una solucion
-		virtual int getAptitude() = 0;
+		virtual double getAptitude() = 0;
 
 	//Estas funciones solo son accesibles desde dentro de la clase o de clases que hereden
 	protected:
